Update robot.cpp
Added in code for identifying when the white path is no longer in the center of the camera view, and changing velocity so that robot moves with the white path. 
Need to work out how to append index of the 4 white pixels of the white path into an array for it to work. 
Need to figure out the 'd' constant, its just a placeholder value for now.