Update robot.cpp
Added 90 degree turn scanning and fast turning for the 90 degree corners in the completion maze. Also added a turn around statement for when the robot detects no white pixels. Added another layer of scanning for when there are gaps in the white path to prevent the robot from turning around due to a minor gap. The robot manages to complete the completion maze, but has too many unnecessary fast turnings and needs to follow the path more closely. Still need to split the code into functions and separate files, and work on the challenge maze.