Update robot.cpp
Added in three camera views for detecting red pixels/walls and three vectors to store the red pixel indexes. Added logic statements for the robot to be able to react to and avoid red walls and complete the challenge maze. Now need to make the red pixel vectors only store red pixels and not also white pixels (this will stop the challenge logic from interfering with the core/completion logic). Need to split the code up into functions with arguments and then into separate files.