Update robot.cpp
I tried to read lines from the config file in order to determine if the program would run based on red pixels or white pixels (challenge or core/completion), but that caused a few errors so instead I added a user input that would determine which methods needed to be called. I might have to add some error checking later or potentially look at reading the config.txt lines again if I have the time, but this works fine for now.